:art: Move `SummaryConfigDTO`
